Begin porting to 1.21.3
Begin porting to 1.21.3.

No runtime testing yet. It just compiles.

+++ src/main/java/tech/anonymoushacker1279/immersiveweapons/item/gun/data/GunData.java
@@ -4,7 +4,7 @@ package tech.anonymoushacker1279.immersiveweapons.item.gun.data;
  * Client-only data used when rendering scope overlays.
  */
 public class GunData {
-	public static double playerFOV = 70.0d;
-	public static double changingPlayerFOV = -1;
+	public static float playerFOV = 70.0f;
+	public static float changingPlayerFOV = -1;
 	public static float scopeScale = 0.5f;
 }
